+++ models/linearElasticity.ts
@@ -73,21 +73,24 @@ const LinearElasticity: IModel = {
           name: 'Rho',
           htmlEntity: 'formula',
           default: '8050',
-          unit: '\\(kg.m^{-3}\\)'
+          units: [{ label: '\\(kg.m^{-3}\\)', multiplicator: 1 }],
+          unit: { label: '\\(kg.m^{-3}\\)', multiplicator: 1 }
         },
         {
           label: "Young's modulus",
           name: 'E',
           htmlEntity: 'formula',
           default: '1e9',
-          unit: '\\(Pa\\)'
+          units: [{ label: '\\(Pa\\)', multiplicator: 1 }],
+          unit: { label: '\\(Pa\\)', multiplicator: 1 }
         },
         {
           label: "Poisson's ratio",
           name: 'Nu',
           htmlEntity: 'formula',
           default: '0.4',
-          unit: '\\(1\\)'
+          units: [{ label: '1', multiplicator: 1 }],
+          unit: { label: '1', multiplicator: 1 }
         }
       ]
     },
@@ -101,20 +104,23 @@ const LinearElasticity: IModel = {
             label: 'External force (x)',
             htmlEntity: 'formula',
             default: '0',
-            unit: '\\(N.m^{-3}\\)'
+            units: [{ label: '\\(N.m^{-3}\\)', multiplicator: 1 }],
+            unit: { label: '\\(N.m^{-3}\\)', multiplicator: 1 }
           },
           {
             label: 'External force (y)',
             htmlEntity: 'formula',
             default: '0',
-            unit: '\\(N.m^{-3}\\)'
+            units: [{ label: '\\(N.m^{-3}\\)', multiplicator: 1 }],
+            unit: { label: '\\(N.m^{-3}\\)', multiplicator: 1 }
           },
           {
             only3D: true,
             label: 'External force (z)',
             htmlEntity: 'formula',
             default: '0',
-            unit: '\\(N.m^{-3}\\)'
+            units: [{ label: '\\(N.m^{-3}\\)', multiplicator: 1 }],
+            unit: { label: '\\(N.m^{-3}\\)', multiplicator: 1 }
           }
         ]
       },
@@ -126,7 +132,8 @@ const LinearElasticity: IModel = {
             label: 'Standard gravity',
             htmlEntity: 'formula',
             default: '9.81',
-            unit: '\\(m.s^{-2}\\)'
+            units: [{ label: '\\(m.s^{-2}\\)', multiplicator: 1 }],
+            unit: { label: '\\(m.s^{-2}\\)', multiplicator: 1 }
           }
         ]
       },
@@ -202,20 +209,23 @@ const LinearElasticity: IModel = {
             label: 'Ux',
             htmlEntity: 'formula',
             default: '0',
-            unit: '\\(m\\)'
+            units: [{ label: '\\(m\\)', multiplicator: 1 }],
+            unit: { label: '\\(m\\)', multiplicator: 1 }
           },
           {
             label: 'Uy',
             htmlEntity: 'formula',
             default: '0',
-            unit: '\\(m\\)'
+            units: [{ label: '\\(m\\)', multiplicator: 1 }],
+            unit: { label: '\\(m\\)', multiplicator: 1 }
           },
           {
             only3D: true,
             label: 'Uz',
             htmlEntity: 'formula',
             default: '0',
-            unit: '\\(m\\)'
+            units: [{ label: '\\(m\\)', multiplicator: 1 }],
+            unit: { label: '\\(m\\)', multiplicator: 1 }
           }
         ],
         refineFactor: 2
@@ -227,7 +237,8 @@ const LinearElasticity: IModel = {
             label: 'd(U)/d(N)',
             htmlEntity: 'formula',
             default: '0',
-            unit: '\\(N.m^{-2}\\)'
+            units: [{ label: '\\(N.m^{-2}\\)', multiplicator: 1 }],
+            unit: { label: '\\(N.m^{-2}\\)', multiplicator: 1 }
           }
         ],
         refineFactor: 2
@@ -239,19 +250,22 @@ const LinearElasticity: IModel = {
             label: 'x',
             htmlEntity: 'formula',
             default: '0',
-            unit: '\\(N.m^{-2}\\)'
+            units: [{ label: '\\(N.m^{-2}\\)', multiplicator: 1 }],
+            unit: { label: '\\(N.m^{-2}\\)', multiplicator: 1 }
           },
           {
             label: 'y',
             htmlEntity: 'formula',
             default: '0',
-            unit: '\\(N.m^{-2}\\)'
+            units: [{ label: '\\(N.m^{-2}\\)', multiplicator: 1 }],
+            unit: { label: '\\(N.m^{-2}\\)', multiplicator: 1 }
           },
           {
             label: 'z',
             htmlEntity: 'formula',
             default: '0',
-            unit: '\\(N.m^{-2}\\)'
+            units: [{ label: '\\(N.m^{-2}\\)', multiplicator: 1 }],
+            unit: { label: '\\(N.m^{-2}\\)', multiplicator: 1 }
           }
         ]
       }