+++ models/stokes.ts
@@ -57,7 +57,8 @@ const Stokes: IModel = {
           name: 'Mu',
           htmlEntity: 'formula',
           default: '1e-3',
-          unit: '\\(Pa.s\\)'
+          units: [{ label: '\\(Pa.s\\)', multiplicator: 1 }],
+          unit: { label: '\\(Pa.s\\)', multiplicator: 1 }
         }
       ]
     },
@@ -71,20 +72,23 @@ const Stokes: IModel = {
             label: 'External force (x)',
             htmlEntity: 'formula',
             default: '0',
-            unit: '\\(N.m^{-3}\\)'
+            units: [{ label: '\\(N.m^{-3}\\)', multiplicator: 1 }],
+            unit: { label: '\\(N.m^{-3}\\)', multiplicator: 1 }
           },
           {
             label: 'External force (y)',
             htmlEntity: 'formula',
             default: '0',
-            unit: '\\(N.m^{-3}\\)'
+            units: [{ label: '\\(N.m^{-3}\\)', multiplicator: 1 }],
+            unit: { label: '\\(N.m^{-3}\\)', multiplicator: 1 }
           },
           {
             only3D: true,
             label: 'External force (z)',
             htmlEntity: 'formula',
             default: '0',
-            unit: '\\(N.m^{-3}\\)'
+            units: [{ label: '\\(N.m^{-3}\\)', multiplicator: 1 }],
+            unit: { label: '\\(N.m^{-3}\\)', multiplicator: 1 }
           }
         ]
       },
@@ -157,20 +161,23 @@ const Stokes: IModel = {
             label: 'Ux',
             htmlEntity: 'formula',
             default: '0',
-            unit: '\\(m.s^{-1}\\)'
+            units: [{ label: '\\(m.s^{-1}\\)', multiplicator: 1 }],
+            unit: { label: '\\(m.s^{-1}\\)', multiplicator: 1 }
           },
           {
             label: 'Uy',
             htmlEntity: 'formula',
             default: '0',
-            unit: '\\(m.s^{-1}\\)'
+            units: [{ label: '\\(m.s^{-1}\\)', multiplicator: 1 }],
+            unit: { label: '\\(m.s^{-1}\\)', multiplicator: 1 }
           },
           {
             only3D: true,
             label: 'Uz',
             htmlEntity: 'formula',
             default: '0',
-            unit: '\\(m.s^{-1}\\)'
+            units: [{ label: '\\(m.s^{-1}\\)', multiplicator: 1 }],
+            unit: { label: '\\(m.s^{-1}\\)', multiplicator: 1 }
           }
         ],
         refineFactor: 2
@@ -182,7 +189,8 @@ const Stokes: IModel = {
             label: 'd(U)/d(N)',
             htmlEntity: 'formula',
             default: '0',
-            unit: '\\(N.m^{-2}\\)'
+            units: [{ label: '\\(N.m^{-2}\\)', multiplicator: 1 }],
+            unit: { label: '\\(N.m^{-2}\\)', multiplicator: 1 }
           }
         ],
         refineFactor: 2